+++ sumologic/resource_sumologic_monitors_library_monitor.go
@@ -280,7 +280,7 @@ func resourceSumologicMonitorsLibraryMonitor() *schema.Resource {
 			"description": {
 				Type:         schema.TypeString,
 				Optional:     true,
-				ValidateFunc: validation.StringMatch(regexp.MustCompile(`^[^\ ].*[^\ ]$`), "description must not contain leading or trailing spaces"),
+				ValidateFunc: validation.StringMatch(regexp.MustCompile(`(?s)^[^\ ].*[^\ ]$`), "description must not contain leading or trailing spaces"),
 			},
 
 			"created_at": {
@@ -337,7 +337,7 @@ func resourceSumologicMonitorsLibraryMonitor() *schema.Resource {
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validation.StringMatch(regexp.MustCompile(`^[^\ ].*[^\ ]$`), "name must not contain leading or trailing spaces"),
+				ValidateFunc: validation.StringMatch(regexp.MustCompile(`(?s)^[^\ ].*[^\ ]$`), "name must not contain leading or trailing spaces"),
 			},
 
 			"post_request_map": {